requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Machine Learning and AI
torch==2.1.2
transformers==4.37.0
scikit-learn==1.4.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP and API
httpx==0.26.0
requests==2.31.0

# Logging and Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Environment
python-dotenv==1.0.0
